ADD Release snapshot without tests

+++ recordstore-service/src/main/java/org/eea/recordstore/service/impl/JdbcRecordStoreServiceImpl.java
@@ -58,7 +58,7 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
   /**
    * The Constant FILE_PATTERN_NAME.
    */
-  private static final String FILE_PATTERN_NAME = "snapshot_%s-dataset_%s%s";
+  private static final String FILE_PATTERN_NAME = "snapshot_%s%s";
 
   /**
    * The kafka sender helper.
@@ -293,15 +293,15 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
       CopyManager cm = new CopyManager((BaseConnection) con);
 
       // Copy dataset_value
-      String nameFileDatasetValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-          idReportingDataset, "_table_DatasetValue.snap");
+      String nameFileDatasetValue =
+          pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_DatasetValue.snap");
       String copyQueryDataset = "COPY (SELECT id, id_dataset_schema FROM dataset_"
           + idReportingDataset + ".dataset_value) to STDOUT";
 
       printToFile(nameFileDatasetValue, copyQueryDataset, cm);
       // Copy table_value
-      String nameFileTableValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-          idReportingDataset, "_table_TableValue.snap");
+      String nameFileTableValue =
+          pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_TableValue.snap");
 
       String copyQueryTable = "COPY (SELECT id, id_table_schema, dataset_id FROM dataset_"
           + idReportingDataset + ".table_value) to STDOUT";
@@ -309,18 +309,18 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
       printToFile(nameFileTableValue, copyQueryTable, cm);
 
       // Copy record_value
-      String nameFileRecordValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-          idReportingDataset, "_table_RecordValue.snap");
+      String nameFileRecordValue =
+          pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_RecordValue.snap");
       String copyQueryRecord =
-          "COPY (SELECT id, id_record_schema, id_table, dataset_partition_id FROM dataset_"
+          "COPY (SELECT id, id_record_schema, id_table, dataset_partition_id,data_provider_code FROM dataset_"
               + idReportingDataset + ".record_value WHERE dataset_partition_id="
               + idPartitionDataset + ") to STDOUT";
 
       printToFile(nameFileRecordValue, copyQueryRecord, cm);
 
       // Copy field_value
-      String nameFileFieldValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-          idReportingDataset, "_table_FieldValue.snap");
+      String nameFileFieldValue =
+          pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_FieldValue.snap");
       String copyQueryField =
           "COPY (SELECT fv.id, fv.type, fv.value, fv.id_field_schema, fv.id_record from dataset_"
               + idReportingDataset + ".field_value fv inner join dataset_" + idReportingDataset
@@ -364,6 +364,7 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
     }
   }
 
+
   /**
    * Restore data snapshot.
    *
@@ -371,14 +372,17 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
    * @param idSnapshot the id snapshot
    * @param partitionId the partition id
    * @param datasetType the dataset type
-   * @param user the user
+   * @param isSchemaSnapshot the is schema snapshot
+   * @param deleteData the delete data
    * @throws SQLException the SQL exception
    * @throws IOException Signals that an I/O exception has occurred.
    */
   @Override
   @Async
   public void restoreDataSnapshot(Long idReportingDataset, Long idSnapshot, Long partitionId,
-      TypeDatasetEnum datasetType, Boolean isSchemaSnapshot) throws SQLException, IOException {
+      TypeDatasetEnum datasetType, Boolean isSchemaSnapshot, Boolean deleteData)
+      throws SQLException, IOException {
+
 
     EventType successEventType =
         isSchemaSnapshot ? EventType.RESTORE_DATASET_SCHEMA_SNAPSHOT_COMPLETED_EVENT
@@ -400,21 +404,22 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
       con = DriverManager.getConnection(conexion.getConnectionString(), conexion.getUser(),
           conexion.getPassword());
       con.setAutoCommit(false);
-      String sql = "";
 
-      switch (datasetType) {
-        case REPORTING:
-          sql = "DELETE FROM dataset_" + idReportingDataset
-              + ".record_value WHERE dataset_partition_id=" + partitionId;
-          break;
-        case DESIGN:
-          sql = "DELETE FROM dataset_" + idReportingDataset + ".table_value";
-          break;
+      if (deleteData) {
+        String sql = "";
+        switch (datasetType) {
+          case REPORTING:
+            sql = "DELETE FROM dataset_" + idReportingDataset
+                + ".record_value WHERE dataset_partition_id=" + partitionId;
+            break;
+          case DESIGN:
+            sql = "DELETE FROM dataset_" + idReportingDataset + ".table_value";
+            break;
+        }
+        stmt = con.createStatement();
+        LOG.info("Deleting previous data");
+        stmt.executeUpdate(sql);
       }
-      stmt = con.createStatement();
-      LOG.info("Deleting previous data");
-      stmt.executeUpdate(sql);
-
 
       CopyManager cm = new CopyManager((BaseConnection) con);
       LOG.info("Init restoring the snapshot files from Snapshot {}", idSnapshot);
@@ -422,8 +427,8 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
         case DESIGN:
           // If it is a design dataset (schema), we need to restore the table values. Otherwise it's
           // not neccesary
-          String nameFileTableValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-              idReportingDataset, "_table_TableValue.snap");
+          String nameFileTableValue =
+              pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_TableValue.snap");
 
           String copyQueryTable = "COPY dataset_" + idReportingDataset
               + ".table_value(id, id_table_schema, dataset_id) FROM STDIN";
@@ -431,16 +436,16 @@ public class JdbcRecordStoreServiceImpl implements RecordStoreService {
           break;
       }
       // Record value
-      String nameFileRecordValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-          idReportingDataset, "_table_RecordValue.snap");
+      String nameFileRecordValue =
+          pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_RecordValue.snap");
 
       String copyQueryRecord = "COPY dataset_" + idReportingDataset
-          + ".record_value(id, id_record_schema, id_table, dataset_partition_id) FROM STDIN";
+          + ".record_value(id, id_record_schema, id_table, dataset_partition_id, data_provider_code) FROM STDIN";
       copyFromFile(copyQueryRecord, nameFileRecordValue, cm);
 
       // Field value
-      String nameFileFieldValue = pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot,
-          idReportingDataset, "_table_FieldValue.snap");
+      String nameFileFieldValue =
+          pathSnapshot + String.format(FILE_PATTERN_NAME, idSnapshot, "_table_FieldValue.snap");
 
       String copyQueryField = "COPY dataset_" + idReportingDataset
           + ".field_value(id, type, value, id_field_schema, id_record) FROM STDIN";
